+++ lib/roadie.rb
@@ -17,6 +17,7 @@ require 'roadie/provider_list'
 require 'roadie/filesystem_provider'
 require 'roadie/null_provider'
 require 'roadie/net_http_provider'
+require 'roadie/cached_provider'
 
 require 'roadie/asset_scanner'
 require 'roadie/markup_improver'